requirements.txt
langchain>=0.2.0
streamlit==1.31.1
faiss-cpu==1.7.4
mlflow==2.11.1
python-dotenv==1.0.1
fpdf==1.7.2
matplotlib==3.8.3
PyPDF2==3.0.1
pdfplumber==0.10.3
langchain-community>=0.0.24
langchain-openai>=0.1.6
pypdf==5.4.0
pytest
langchain==0.3.25
llama-cloud==0.1.22
llama-cloud-services==0.6.23
llama-index==0.12.37
llama-index-agent-openai==0.4.8
llama-index-cli==0.4.1
llama-index-core==0.12.37
llama-index-embeddings-ollama==0.6.0
llama-index-embeddings-openai==0.3.1
llama-index-indices-managed-llama-cloud==0.6.11
llama-index-llms-ollama==0.5.5
llama-index-llms-openai==0.3.43
llama-index-multi-modal-llms-openai==0.4.3
llama-index-program-openai==0.3.1
llama-index-question-gen-openai==0.3.0
llama-index-readers-file==0.4.7
llama-index-readers-llama-parse==0.4.0
llama-parse==0.6.23
